+++ Changelog.md
@@ -39,3 +39,4 @@
 - Removed "showErrors" config option. A warning/error message will always display. However, you can disable the java errors with "showErrorTrace".
 - If a plugin version has multiple files it will prompt you for which ones to install.
 - Added "/plugin reload", which makes CogWorks rescan the "./plugins" folder for any changes.
+- If a plugin doesn't have a config folder then there won't be a prompt to delete the config folder.